+++ cedar-policy-core/src/ast/expr.rs
@@ -144,6 +144,14 @@ pub enum ExprKind<T = ()> {
         /// Be careful the backslash in `\*` must not be another escape sequence. For instance, `\\*` matches a backslash plus an arbitrary string.
         pattern: Pattern,
     },
+    /// Entity type test. Does the first argument have the entity type
+    /// specified by the second argument.
+    Is {
+        /// Expression to test. Must evaluate to an Entity.
+        expr: Arc<Expr<T>>,
+        /// The entity type `Name` used for the type membership test.
+        entity_type: Name,
+    },
     /// Set (whose elements may be arbitrary expressions)
     //
     // This is backed by `Vec` (and not e.g. `HashSet`), because two `Expr`s
@@ -465,6 +473,11 @@ impl Expr {
         ExprBuilder::new().like(expr, pattern)
     }
 
+    /// Create an `is` expression.
+    pub fn is_entity_type(expr: Expr, entity_type: Name) -> Self {
+        ExprBuilder::new().is_entity_type(expr, entity_type)
+    }
+
     /// Check if an expression contains any symbolic unknowns
     pub fn is_unknown(&self) -> bool {
         self.subexpressions()
@@ -568,6 +581,10 @@ impl Expr {
             ExprKind::MulByConst { arg, constant } => {
                 Ok(Expr::mul(arg.substitute(definitions)?, *constant))
             }
+            ExprKind::Is { expr, entity_type } => Ok(Expr::is_entity_type(
+                expr.substitute(definitions)?,
+                entity_type.clone(),
+            )),
         }
     }
 }
@@ -949,6 +966,14 @@ impl<T> ExprBuilder<T> {
             pattern: Pattern::new(pattern),
         })
     }
+
+    /// Create an 'is' expression.
+    pub fn is_entity_type(self, expr: Expr<T>, entity_type: Name) -> Expr<T> {
+        self.with_expr_kind(ExprKind::Is {
+            expr: Arc::new(expr),
+            entity_type,
+        })
+    }
 }
 
 impl<T: Clone> ExprBuilder<T> {
@@ -1149,6 +1174,13 @@ impl<T> Expr<T> {
                         .zip(map1.iter()) // relying on BTreeMap producing an iterator sorted by key
                         .all(|((a, e), (a1, e1))| a == a1 && e.eq_shape(e1))
             }
+            (
+                Is { expr, entity_type },
+                Is {
+                    expr: expr1,
+                    entity_type: entity_type1,
+                },
+            ) => entity_type == entity_type1 && expr.eq_shape(expr1),
             _ => false,
         }
     }
@@ -1233,6 +1265,10 @@ impl<T> Expr<T> {
                     s.hash(state);
                     a.hash_shape(state);
                 });
+            }
+            ExprKind::Is { expr, entity_type } => {
+                expr.hash_shape(state);
+                entity_type.hash(state);
             }
         }
     }
@@ -1392,6 +1428,24 @@ mod test {
                 None,
                 ()
             )
+        );
+        assert_eq!(
+            Expr::is_entity_type(
+                Expr::val(EntityUID::with_eid("foo")),
+                "Type".parse().unwrap()
+            ),
+            Expr::new(
+                ExprKind::Is {
+                    expr: Arc::new(Expr::new(
+                        ExprKind::Lit(Literal::from(EntityUID::with_eid("foo"))),
+                        None,
+                        ()
+                    )),
+                    entity_type: "Type".parse().unwrap()
+                },
+                None,
+                ()
+            ),
         );
     }
 
@@ -1601,8 +1655,12 @@ mod test {
                 Expr::has_attr(Expr::val(1), "foo".into()),
             ),
             (
-                ExprBuilder::with_data(1).like(temp, vec![PatternElem::Wildcard]),
+                ExprBuilder::with_data(1).like(temp.clone(), vec![PatternElem::Wildcard]),
                 Expr::like(Expr::val(1), vec![PatternElem::Wildcard]),
+            ),
+            (
+                ExprBuilder::with_data(1).is_entity_type(temp, "T".parse().unwrap()),
+                Expr::is_entity_type(Expr::val(1), "T".parse().unwrap()),
             ),
         ];