+++ cedar-policy-core/src/ast/restricted_expr.rs
@@ -325,6 +325,10 @@ fn is_restricted(expr: &Expr) -> Result<(), RestrictedExprError> {
             feature: "'like'".into(),
             expr: expr.clone(),
         }),
+        ExprKind::Is { .. } => Err(RestrictedExprError::InvalidRestrictedExpression {
+            feature: "'is'".into(),
+            expr: expr.clone(),
+        }),
         ExprKind::ExtensionFunctionApp { args, .. } => args.iter().try_for_each(is_restricted),
         ExprKind::Set(exprs) => exprs.iter().try_for_each(is_restricted),
         ExprKind::Record(map) => map.values().try_for_each(is_restricted),